Remove bugs from update functionality and added Delete functionality.

+++ src/main/java/ExpenseTracker/model/Expense.java
@@ -2,8 +2,6 @@ package ExpenseTracker.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
 
 @Data
 
@@ -11,14 +9,14 @@ import lombok.ToString;
 public class Expense {
     private int id;
     private String description;
-    private double amount;
+    private int amount;
 
     public String toCsv() {
         return id + "," + description + "," + amount + "\n";
     }
     public static Expense fromCsv(String csv) {
         String[] values = csv.split(",");
-        return new Expense(Integer.parseInt(values[0]),values[1],Double.parseDouble(values[2]));
+        return new Expense(Integer.parseInt(values[0]),values[1],Integer.parseInt(values[2]));
     }
     public String toString() {
         return "id:" +id + ",Description:"+ description + ", Amount:" + amount ;
